profile_prod/validacion/requirements.txt
# FastAPI and ASGI server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Deep Learning
torch==2.1.0
torchvision==0.16.0
torchaudio==2.1.0

# Computer Vision
opencv-python==4.8.1.78
Pillow==10.1.0

# Data manipulation
numpy==1.24.3
pandas==2.1.3

# Visualization
matplotlib==3.8.2
seaborn==0.13.0

# HTTP client (for health checks)
requests==2.31.0

# Logging and utilities
pydantic==2.5.0
python-jose[cryptography]==3.3.0

# Additional utilities
pathlib2==2.3.7
